fix(ui): revert changes (#245)
