+++ ui/src/client/api/invitation.ts
@@ -8,11 +8,20 @@
 // by the GNU Affero General Public License v3.0 only, included in the file
 // licenses/AGPL.txt.
 import useSWR, { SWRConfiguration } from 'swr'
-import { paramsFromListOptions } from '@/client/api/query-helpers'
-import { ListOptions } from '@/client/api/types/queries'
 import { apiFetcher } from '@/client/fetcher'
 import { User } from '@/client/idp/user'
 import { Organization } from './organization'
+
+export enum SortBy {
+  Email = 'email',
+  DateCreated = 'date_created',
+  DateModified = 'date_modified',
+}
+
+export enum SortOrder {
+  Asc = 'asc',
+  Desc = 'desc',
+}
 
 export type InvitationStatus = 'pending' | 'accepted' | 'declined'
 
@@ -39,6 +48,23 @@ export type CreateOptions = {
   emails: string[]
 }
 
+export type ListOptions = {
+  organizationId?: string
+  size?: number
+  page?: number
+  sortBy?: SortBy
+  sortOrder?: SortOrder
+}
+
+type ListQueryParams = {
+  page?: string
+  size?: string
+  sort_by?: string
+  sort_order?: string
+  query?: string
+  organization_id?: string
+}
+
 export default class InvitationAPI {
   static async create(options: CreateOptions) {
     return apiFetcher({
@@ -49,7 +75,7 @@ export default class InvitationAPI {
   }
 
   static useGetIncoming(options?: ListOptions, swrOptions?: SWRConfiguration) {
-    const url = `/invitations/incoming?${paramsFromListOptions(options)}`
+    const url = `/invitations/incoming?${this.paramsFromListOptions(options)}`
     return useSWR<List>(
       url,
       () => apiFetcher({ url, method: 'GET' }) as Promise<List>,
@@ -67,12 +93,32 @@ export default class InvitationAPI {
   }
 
   static useGetOutgoing(options?: ListOptions, swrOptions?: SWRConfiguration) {
-    const url = `/invitations/outgoing?${paramsFromListOptions(options)}`
+    const url = `/invitations/outgoing?${this.paramsFromListOptions(options)}`
     return useSWR<List>(
       options?.organizationId ? url : null,
       () => apiFetcher({ url, method: 'GET' }) as Promise<List>,
       swrOptions,
     )
+  }
+
+  static paramsFromListOptions(options?: ListOptions): URLSearchParams {
+    const params: ListQueryParams = {}
+    if (options?.organizationId) {
+      params.organization_id = options.organizationId.toString()
+    }
+    if (options?.page) {
+      params.page = options.page.toString()
+    }
+    if (options?.size) {
+      params.size = options.size.toString()
+    }
+    if (options?.sortBy) {
+      params.sort_by = options.sortBy.toString()
+    }
+    if (options?.sortOrder) {
+      params.sort_order = options.sortOrder.toString()
+    }
+    return new URLSearchParams(params)
   }
 
   static async delete(id: string) {
