+++ ui/src/client/api/organization.ts
@@ -8,10 +8,19 @@
 // by the GNU Affero General Public License v3.0 only, included in the file
 // licenses/AGPL.txt.
 import useSWR, { SWRConfiguration } from 'swr'
-import { paramsFromListOptions } from '@/client/api/query-helpers'
-import { ListOptions } from '@/client/api/types/queries'
 import { apiFetcher } from '@/client/fetcher'
 import { PermissionType } from './permission'
+
+export enum SortBy {
+  Name = 'name',
+  DateCreated = 'date_created',
+  DateModified = 'date_modified',
+}
+
+export enum SortOrder {
+  Asc = 'asc',
+  Desc = 'desc',
+}
 
 export type Organization = {
   id: string
@@ -29,6 +38,14 @@ export type List = {
   size: number
 }
 
+export type ListOptions = {
+  query?: string
+  size?: number
+  page?: number
+  sortBy?: SortBy
+  sortOrder?: SortOrder
+}
+
 export type CreateOptions = {
   name: string
   image?: string
@@ -40,6 +57,14 @@ export type PatchNameOptions = {
 
 export type RemoveMemberOptions = {
   userId: string
+}
+
+type ListQueryParams = {
+  page?: string
+  size?: string
+  sort_by?: string
+  sort_order?: string
+  query?: string
 }
 
 export default class OrganizationAPI {
@@ -54,18 +79,38 @@ export default class OrganizationAPI {
 
   static async list(options?: ListOptions) {
     return apiFetcher({
-      url: `/organizations?${paramsFromListOptions(options)}`,
+      url: `/organizations?${this.paramsFromListOptions(options)}`,
       method: 'GET',
     }) as Promise<List>
   }
 
   static useList(options?: ListOptions, swrOptions?: SWRConfiguration) {
-    const url = `/organizations?${paramsFromListOptions(options)}`
+    const url = `/organizations?${this.paramsFromListOptions(options)}`
     return useSWR<List>(
       url,
       () => apiFetcher({ url, method: 'GET' }) as Promise<List>,
       swrOptions,
     )
+  }
+
+  static paramsFromListOptions(options?: ListOptions): URLSearchParams {
+    const params: ListQueryParams = {}
+    if (options?.query) {
+      params.query = encodeURIComponent(options.query.toString())
+    }
+    if (options?.page) {
+      params.page = options.page.toString()
+    }
+    if (options?.size) {
+      params.size = options.size.toString()
+    }
+    if (options?.sortBy) {
+      params.sort_by = options.sortBy.toString()
+    }
+    if (options?.sortOrder) {
+      params.sort_order = options.sortOrder.toString()
+    }
+    return new URLSearchParams(params)
   }
 
   static async create(options: CreateOptions) {
