+++ ui/src/client/api/user.ts
@@ -8,9 +8,17 @@
 // by the GNU Affero General Public License v3.0 only, included in the file
 // licenses/AGPL.txt.
 import useSWR, { SWRConfiguration } from 'swr'
-import { paramsFromListOptions } from '@/client/api/query-helpers'
-import { ListOptions } from '@/client/api/types/queries'
 import { apiFetcher } from '@/client/fetcher'
+
+export enum SortBy {
+  Email = 'email',
+  FullName = 'full_name',
+}
+
+export enum SortOrder {
+  Asc = 'asc',
+  Desc = 'desc',
+}
 
 export type User = {
   id: string
@@ -28,20 +36,71 @@ export type List = {
   size: number
 }
 
+export type ListOptions = {
+  query?: string
+  organizationId?: string
+  groupId?: string
+  excludeGroupMembers?: boolean
+  size?: number
+  page?: number
+  sortBy?: SortBy
+  sortOrder?: SortOrder
+}
+
+type ListQueryParams = {
+  page?: string
+  size?: string
+  sort_by?: string
+  sort_order?: string
+  query?: string
+  organization_id?: string
+  group_id?: string
+  exclude_group_members?: string
+}
+
 export default class UserAPI {
   static async list(options?: ListOptions) {
     return apiFetcher({
-      url: `/users?${paramsFromListOptions(options)}`,
+      url: `/users?${this.paramsFromListOptions(options)}`,
       method: 'GET',
     }) as Promise<List>
   }
 
   static useList(options?: ListOptions, swrOptions?: SWRConfiguration) {
-    const url = `/users?${paramsFromListOptions(options)}`
+    const url = `/users?${this.paramsFromListOptions(options)}`
     return useSWR<List>(
       url,
       () => apiFetcher({ url, method: 'GET' }) as Promise<List>,
       swrOptions,
     )
+  }
+
+  static paramsFromListOptions(options?: ListOptions): URLSearchParams {
+    const params: ListQueryParams = {}
+    if (options?.query) {
+      params.query = encodeURIComponent(options.query.toString())
+    }
+    if (options?.organizationId) {
+      params.organization_id = options.organizationId.toString()
+    }
+    if (options?.groupId) {
+      params.group_id = options.groupId.toString()
+    }
+    if (options?.excludeGroupMembers) {
+      params.exclude_group_members = options.excludeGroupMembers.toString()
+    }
+    if (options?.page) {
+      params.page = options.page.toString()
+    }
+    if (options?.size) {
+      params.size = options.size.toString()
+    }
+    if (options?.sortBy) {
+      params.sort_by = options.sortBy.toString()
+    }
+    if (options?.sortOrder) {
+      params.sort_order = options.sortOrder.toString()
+    }
+    return new URLSearchParams(params)
   }
 }
