+++ ui/src/client/api/workspace.ts
@@ -8,11 +8,20 @@
 // by the GNU Affero General Public License v3.0 only, included in the file
 // licenses/AGPL.txt.
 import useSWR, { SWRConfiguration } from 'swr'
-import { paramsFromListOptions } from '@/client/api/query-helpers'
-import { ListOptions } from '@/client/api/types/queries'
 import { apiFetcher } from '@/client/fetcher'
 import { Organization } from './organization'
 import { PermissionType } from './permission'
+
+export enum SortBy {
+  Name = 'name',
+  DateCreated = 'date_created',
+  DateModified = 'date_modified',
+}
+
+export enum SortOrder {
+  Asc = 'asc',
+  Desc = 'desc',
+}
 
 export type Workspace = {
   id: string
@@ -40,12 +49,28 @@ export interface CreateOptions {
   storageCapacity: number
 }
 
+export type ListOptions = {
+  query?: string
+  size?: number
+  page?: number
+  sortBy?: SortBy
+  sortOrder?: SortOrder
+}
+
 export interface PatchNameOptions {
   name: string
 }
 
 export interface PatchStorageCapacityOptions {
   storageCapacity: number
+}
+
+type ListQueryParams = {
+  page?: string
+  size?: string
+  sort_by?: string
+  sort_order?: string
+  query?: string
 }
 
 export default class WorkspaceAPI {
@@ -59,7 +84,7 @@ export default class WorkspaceAPI {
   }
 
   static useList(options?: ListOptions, swrOptions?: SWRConfiguration) {
-    const url = `/workspaces?${paramsFromListOptions(options)}`
+    const url = `/workspaces?${this.paramsFromListOptions(options)}`
     return useSWR<List>(
       url,
       () => apiFetcher({ url, method: 'GET' }) as Promise<List>,
@@ -69,9 +94,29 @@ export default class WorkspaceAPI {
 
   static async list(options?: ListOptions) {
     return apiFetcher({
-      url: `/workspaces?${paramsFromListOptions(options)}`,
+      url: `/workspaces?${this.paramsFromListOptions(options)}`,
       method: 'GET',
     }) as Promise<List>
+  }
+
+  static paramsFromListOptions(options?: ListOptions): URLSearchParams {
+    const params: ListQueryParams = {}
+    if (options?.query) {
+      params.query = encodeURIComponent(options.query.toString())
+    }
+    if (options?.page) {
+      params.page = options.page.toString()
+    }
+    if (options?.size) {
+      params.size = options.size.toString()
+    }
+    if (options?.sortBy) {
+      params.sort_by = options.sortBy.toString()
+    }
+    if (options?.sortOrder) {
+      params.sort_order = options.sortOrder.toString()
+    }
+    return new URLSearchParams(params)
   }
 
   static async create(options: CreateOptions) {
