+++ ui/src/components/file/file-browse.tsx
@@ -11,8 +11,7 @@ import { useCallback, useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { Button } from '@chakra-ui/react'
 import cx from 'classnames'
-import FileAPI, { File } from '@/client/api/file'
-import { FileType } from '@/client/api/types/queries'
+import FileAPI, { File, FileType } from '@/client/api/file'
 import WorkspaceAPI from '@/client/api/workspace'
 import Path from '@/components/common/path'
 import FolderSvg from '@/components/file/list/item/icon/icon-folder/assets/icon-folder.svg'
@@ -130,7 +129,6 @@ const FileBrowse = ({ onChange }: FileBrowseProps) => {
             >
               <img
                 src={FolderSvg}
-                alt="Directory"
                 className={cx('shrink-0', 'w-[36px]', 'h-[28.84px]')}
               />
               <Text noOfLines={1}>{f.name}</Text>
