+++ ui/src/components/file/file-toolbar.tsx
@@ -21,9 +21,8 @@ import {
   Tooltip,
 } from '@chakra-ui/react'
 import cx from 'classnames'
-import FileAPI, { List } from '@/client/api/file'
+import FileAPI, { List, SortBy, SortOrder } from '@/client/api/file'
 import { ltEditorPermission } from '@/client/api/permission'
-import { SortBy, SortOrder } from '@/client/api/types/queries'
 import {
   IconAdd,
   IconMoreVert,
