+++ ui/src/components/file/list/list-draggable-droppable.tsx
@@ -17,8 +17,7 @@ import {
   useDroppable,
 } from '@dnd-kit/core'
 import cx from 'classnames'
-import FileAPI from '@/client/api/file'
-import { FileType } from '@/client/api/types/queries'
+import FileAPI, { FileType } from '@/client/api/file'
 import store from '@/store/configure-store'
 import { useAppDispatch, useAppSelector } from '@/store/hook'
 import { hiddenUpdated, selectionUpdated } from '@/store/ui/files'
