+++ ui/src/components/insights/insights-overview/insights-overview-chart.tsx
@@ -11,8 +11,7 @@ import { useMemo } from 'react'
 import { useColorMode } from '@chakra-ui/react'
 import { ResponsivePie } from '@nivo/pie'
 import cx from 'classnames'
-import InsightsAPI from '@/client/api/insights'
-import { SortBy, SortOrder } from '@/client/api/types/queries'
+import InsightsAPI, { SortBy, SortOrder } from '@/client/api/insights'
 import { swrConfig } from '@/client/options'
 import { useAppSelector } from '@/store/hook'
