+++ ui/src/components/insights/insights-overview/insights-overview-entities.tsx
@@ -10,8 +10,7 @@
 import { useCallback, useEffect, useState } from 'react'
 import { Badge, Table, Tbody, Td, Tooltip, Tr } from '@chakra-ui/react'
 import cx from 'classnames'
-import InsightsAPI from '@/client/api/insights'
-import { SortBy, SortOrder } from '@/client/api/types/queries'
+import InsightsAPI, { SortBy, SortOrder } from '@/client/api/insights'
 import { swrConfig } from '@/client/options'
 import Pagination from '@/lib/components/pagination'
 import SearchInput from '@/lib/components/search-input'
