+++ ui/src/components/task/task-list.tsx
@@ -9,8 +9,7 @@
 // licenses/AGPL.txt.
 import { useEffect, useState } from 'react'
 import cx from 'classnames'
-import TaskAPI from '@/client/api/task'
-import { SortOrder } from '@/client/api/types/queries'
+import TaskAPI, { SortOrder } from '@/client/api/task'
 import { swrConfig } from '@/client/options'
 import Pagination from '@/lib/components/pagination'
 import { useAppDispatch } from '@/store/hook'
