+++ ui/src/components/upload/upload-worker.ts
@@ -8,8 +8,7 @@
 // by the GNU Affero General Public License v3.0 only, included in the file
 // licenses/AGPL.txt.
 import { FileWithPath } from 'react-dropzone'
-import FileAPI from '@/client/api/file'
-import { FileType } from '@/client/api/types/queries'
+import FileAPI, { FileType } from '@/client/api/file'
 import { errorToString } from '@/client/error'
 import store from '@/store/configure-store'
 import {
