+++ ui/src/local-storage/index.ts
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the GNU Affero General Public License v3.0 only, included in the file
 // licenses/AGPL.txt.
-import { SortBy, SortOrder } from '@/client/api/types/queries'
+import { SortBy, SortOrder } from '@/client/api/file'
 import { FileViewType } from '@/types/file'
 
 const ICON_SCALE_KEY = 'voltaserve_file_icon_scale'
