+++ ui/src/pages/account/account-invitations-page.tsx
@@ -26,8 +26,7 @@ import {
 } from '@chakra-ui/react'
 import cx from 'classnames'
 import { Helmet } from 'react-helmet-async'
-import InvitationAPI from '@/client/api/invitation'
-import { SortBy, SortOrder } from '@/client/api/types/queries'
+import InvitationAPI, { SortBy, SortOrder } from '@/client/api/invitation'
 import UserAPI from '@/client/idp/user'
 import { swrConfig } from '@/client/options'
 import { incomingInvitationPaginationStorage } from '@/infra/pagination'
