+++ ui/src/pages/group/group-list-page.tsx
@@ -28,8 +28,7 @@ import {
 } from '@chakra-ui/react'
 import cx from 'classnames'
 import { Helmet } from 'react-helmet-async'
-import GroupAPI from '@/client/api/group'
-import { SortOrder } from '@/client/api/types/queries'
+import GroupAPI, { SortOrder } from '@/client/api/group'
 import { swrConfig } from '@/client/options'
 import { CreateGroupButton } from '@/components/app-bar/app-bar-buttons'
 import { groupPaginationStorage } from '@/infra/pagination'
