+++ ui/src/pages/group/group-members-page.tsx
@@ -34,8 +34,7 @@ import cx from 'classnames'
 import { Helmet } from 'react-helmet-async'
 import GroupAPI from '@/client/api/group'
 import { geEditorPermission } from '@/client/api/permission'
-import { SortBy, SortOrder } from '@/client/api/types/queries'
-import UserAPI from '@/client/api/user'
+import UserAPI, { SortBy, SortOrder } from '@/client/api/user'
 import { User as IdPUser } from '@/client/idp/user'
 import { swrConfig } from '@/client/options'
 import GroupAddMember from '@/components/group/group-add-member'
