+++ ui/src/pages/organization/organization-list-page.tsx
@@ -28,8 +28,7 @@ import {
 } from '@chakra-ui/react'
 import cx from 'classnames'
 import { Helmet } from 'react-helmet-async'
-import OrganizationAPI from '@/client/api/organization'
-import { SortOrder } from '@/client/api/types/queries'
+import OrganizationAPI, { SortOrder } from '@/client/api/organization'
 import { swrConfig } from '@/client/options'
 import { CreateOrganizationButton } from '@/components/app-bar/app-bar-buttons'
 import { organizationPaginationStorage } from '@/infra/pagination'
