+++ ui/src/pages/organization/organization-members-page.tsx
@@ -34,8 +34,7 @@ import cx from 'classnames'
 import { Helmet } from 'react-helmet-async'
 import OrganizationAPI from '@/client/api/organization'
 import { geEditorPermission } from '@/client/api/permission'
-import { SortBy, SortOrder } from '@/client/api/types/queries'
-import UserAPI, { User } from '@/client/api/user'
+import UserAPI, { SortBy, SortOrder, User } from '@/client/api/user'
 import { swrConfig } from '@/client/options'
 import OrganizationInviteMembers from '@/components/organization/organization-invite-members'
 import OrganizationRemoveMember from '@/components/organization/organization-remove-member'
