+++ ui/src/pages/workspace/workspace-list-page.tsx
@@ -28,8 +28,7 @@ import {
 } from '@chakra-ui/react'
 import cx from 'classnames'
 import { Helmet } from 'react-helmet-async'
-import { SortOrder } from '@/client/api/types/queries'
-import WorkspaceAPI from '@/client/api/workspace'
+import WorkspaceAPI, { SortOrder } from '@/client/api/workspace'
 import { swrConfig } from '@/client/options'
 import { CreateWorkspaceButton } from '@/components/app-bar/app-bar-buttons'
 import { workspacePaginationStorage } from '@/infra/pagination'
